fix: Tail Detection Emits Boundary Chunk Immediately
Tail detetion no longer holds on to the undefined "boundary" chunk while it waits to detect a tail in subsequent chunks; emits it immediately upon receiving it.

As a result of this fix, a part body is now available to be read as soon as the boundary is detected in the stream. Fixes #1.

Signed-off-by: Rahul Gupta <CxRes@users.noreply.github.com>